fix untested code and use as main vae
