+++ .gitignore
@@ -1,6 +1,7 @@
 # MNIST downloaded data
-MNIST/
-out/
+data
+results/
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
